+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to Hytale Toolkit will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2025-01-25
+
+### Added
+- Database page now shows Hytale version info (fetched from CDN manifest)
+- Shows installed database version when existing database is found
+- Displays update warning when newer database version is available
+
 ## [1.0.4] - 2025-01-24
 
 ### Fixed